Fixed color calculations and using median
Moved circular buffer logic into proper class, so can use it for RGB and
temperature.  Still using average for color, but now using median for
temperature as it seems to be better in face of noise.

Also corrected color picking based on temperture as it was off by one
and interpolating wrong way round.